+++ ricaun.Nuke.PackageBuilder/Components/IHazOutput.cs
@@ -6,6 +6,9 @@ using Nuke.Common.ValueInjection;
 using ricaun.Nuke.Extensions;
 namespace ricaun.Nuke.Components
 {
+    /// <summary>
+    /// IHazOutput
+    /// </summary>
     public interface IHazOutput : IHazPackageBuilderProject, IHazSolution, INukeBuild
     {
         /// <summary>
@@ -19,7 +22,17 @@ namespace ricaun.Nuke.Components
         /// </summary>
         [Parameter]
         public Project Project => ValueInjectionUtility.TryGetValue(() => Project) ?? GetPackageBuilderProject();
+
+        /// <summary>
+        /// OutputDirectory
+        /// </summary>
         AbsolutePath OutputDirectory => GetOutputDirectory(Project);
+
+        /// <summary>
+        /// GetOutputDirectory
+        /// </summary>
+        /// <param name="project"></param>
+        /// <returns></returns>
         public AbsolutePath GetOutputDirectory(Project project) => project.Directory / "bin" / Folder;
     }
 }