+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.0.8] / 2021-12-21
+- Add Documentation on the main project
+- Add `xml` DocumentationFile
+- Force to `ricaun.Nuke` Version - 0.0.12
+- Change `Content` to `Release` on Example
+
 ## [0.0.7] / 2021-12-17
 - Exemple Update Version
 - Add icon.ico
@@ -46,6 +52,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: https://github.com/ricaun-io/ricaun.Nuke/compare/1.0.0...HEAD
+[0.0.8]: https://github.com/ricaun-io/ricaun.Nuke/compare/0.0.7...0.0.8
 [0.0.7]: https://github.com/ricaun-io/ricaun.Nuke/compare/0.0.6...0.0.7
 [0.0.6]: https://github.com/ricaun-io/ricaun.Nuke/compare/0.0.5...0.0.6
 [0.0.5]: https://github.com/ricaun-io/ricaun.Nuke/compare/0.0.4...0.0.5